add validations

+++ src/routes.js
@@ -20,6 +20,10 @@ export const routes = [
     handler: (req, res) => {
       const { title, description } = req.body;
 
+      if (!title || !description) {
+        return res.writeHead(400).end('Necessário informar título e descrição');
+      }
+
       const task = {
         id: randomUUID(),
         title,
@@ -40,8 +44,16 @@ export const routes = [
     handler: (req, res) => {
       const { id } = req.params;
       const { title, description } = req.body;
+      const [task] = database.select('tasks', { id });
+
+      if (!task) return res.writeHead(404).end('Tarefa não encontrada');
+
+      if (!title || !description) {
+        return res.writeHead(400).end('Necessário informar título e descrição');
+      }
 
       database.update('tasks', id, {
+        ...task,
         title,
         description,
         updated_at: new Date(),
@@ -56,6 +68,10 @@ export const routes = [
     handler: (req, res) => {
       const { id } = req.params;
 
+      const [task] = database.select('tasks', { id });
+
+      if (!task) return res.writeHead(404).end('Tarefa não encontrada');
+
       database.delete('tasks', id);
 
       return res.writeHead(204).end();
@@ -67,7 +83,12 @@ export const routes = [
     handler: (req, res) => {
       const { id } = req.params;
 
+      const [task] = database.select('tasks', { id });
+
+      if (!task) return res.writeHead(404).end('Tarefa não encontrada');
+
       database.update('tasks', id, {
+        ...task,
         completed_at: new Date(),
       });
       return res.writeHead(204).end();
